Clean up some Visual Studio level 4 warnings
Some level 4 warnings still remain (notably C4389 - mismatch
of signed int socket handles in Linux and unsigned int socket
handles in Windows

+++ api/src/wsa_api.c
@@ -542,7 +542,7 @@ int16_t wsa_get_decimation(struct wsa_device *dev, int32_t *rate)
 
 	// Handle the query output here 
 	if (query.status <= 0)
-		return (int32_t) query.status;
+		return (int16_t) query.status;
 
 	// convert & make sure no error
 	if (to_int(query.output, &temp) < 0)
